+++ index.html
@@ -234,6 +234,8 @@
 
 
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="util-data.js"></script>
+  <script src="util.js" type="module"></script>
 </body>
 </html>
